+++ apdt/config.py
@@ -1,6 +1,6 @@
 
 _config = {
-  'darksky_secret_key_num': 6,
+  'darksky_secret_key_num': 1,
   'darksky_secret_key_0': '3636efb586ef9346635b4c2481ada25b',
   'nms_data_path': 'data/',
   'foshan_data_path': 'data/foshan_data/',